Move dispatcher to pkg/client/daemon package.
This move is in preparation for merging the dispatcher with the router.
No code change.
